requirements.txt
# ==========================================
# Servidor Web / API (Render - Produção)
# ==========================================
fastapi>=0.115.0
uvicorn>=0.30.0
gunicorn>=23.0.0

# ==========================================
# Inteligência Artificial (Google Gemini)
# ==========================================
google-generativeai>=0.8.6

# ==========================================
# Google Sheets & Autenticação
# ==========================================
gspread>=6.1.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# ==========================================
# Banco de Dados & HTTP
# ==========================================
sqlalchemy>=2.0.0
requests>=2.32.0

# ==========================================
# Timezone & Data/Hora
# ==========================================
pytz>=2023.3

# ==========================================
# Utilitários
# ==========================================
pydantic>=2.7,<3
python-dotenv>=1.0.1